// src/app/_core/domains/enums/action.type.ts
export enum ActionType {
    Empty = '',
    View = 'Xem',
    Edit = 'Sửa',
    Play = 'Chạy',
    Stop = 'Dừng',
    Delete = 'Xóa',
    Cancel = 'Hủy',
    Sync = 'Đồng bộ',
    Copy = 'Sao chép',
    Notes = 'Ghi chú',
    Call = 'Gọi điện',
    Approve = 'Duyệt',
    Reply = 'Trả lời',
    Reject = 'Từ chối',
    Pause = 'Tạm dừng',
    Upload = 'Tải lên',
    Role = 'Phân quyền',
    AddNew = 'Thêm mới',
    Confirm = 'Xác nhận',
    Public = 'Công khai',
    Publish = 'Xuất bản',
    Verify = 'Kiểm duyệt',
    Payment = 'Thanh toán',
    Auditing = 'Kiểm toán',
    PushArticle = 'Đẩy tin',
    IsSupport = 'Nhận CSKH',
    Lock = 'Khóa tài khoản',
    Import = 'Nhập dữ liệu',
    Export = 'Xuất dữ liệu',
    History = 'Xem lịch sử',
    CreateOTP = 'Tạo OTP',
    ViewOTP = 'Xem OTP',
    SendEmail = 'Gửi Email',
    Statistical = "Thống kê",
    AssignSale = 'Điều chuyển',
    CancelArticle = 'Hạ tin',
    GetResult = 'Lấy kết quả',
    Configuration = 'Cấu hình',
    AssignSupport = 'Gán CSKH',
    EditPrice = 'Thay đổi giá',
    AddNewCombo = 'Thêm combo',
    MPDeposit = 'Xem nạp tiền',
    ReceiveArticle = 'Nhận tin',
    ViewDetail = 'Xem chi tiết',
    SyncArticle = 'Đồng bộ tin',
    ApproveArticle = 'Duyệt tin',
    UnLock = 'Mở khóa tài khoản',
    MPPayment = 'Xem thanh toán',
    MPWithdrawl = 'Xem rút tiền',
    ImportToCRM = 'Nhập vào CRM',
    GroupInvoice = "Gộp hóa đơn",
    SplitInvoice = "Tách hóa đơn",
    AddNewSerive = 'Thêm dịch vụ',
    EditMember = 'Thêm nhân viên',
    ResetCache = 'Xóa bộ nhớ tạm',
    CreateMeeyId = 'Tạo tài khoản',
    MPTransfer = 'Xem chuyển tiền',
    VerifyArticle = 'Xác minh tin',
    RePostArticle = 'Đăng lại tin',
    OpenCouponHmc = 'Mở lại Coupon',
    ApproveOrder = 'Duyệt đơn hàng',
    GroupCustomer = 'Gộp khách hàng',
    ApproveRequest = "Duyệt yêu cầu",
    ApproveDeposit = 'Duyệt nạp tiền',
    ApproveContent = 'Duyệt nội dung',
    ApproveWithdraw = 'Duyệt rút tiền',
    AccountantApprove = 'Kế toán duyệt',
    UpdateStatus = 'Chuyển trạng thái',
    TransferArticle = 'Điều chuyển tin',
    AddNewTransaction = "Tạo giao dịch",
    ReceiveCustomer = 'Nhận khách hàng',
    ResetPassword = 'Thiết lập mật khẩu',
    ProcessTransaction = 'Xử lý giao dịch',
    DuplicateArticle = 'Xác minh tin trùng',
    ViewRequestRegister = 'Xem yêu cầu đăng ký',
    UpdateStatusSuccess = 'Chuyển trạng thái thành công',
    ViewDetailInvoiceExported = "Xem chi tiết hóa đơn đã xuất",
    ListenAgain = 'Nghe lại',
    FileDownload = 'Tải file',
    PausePublish = 'Tạm dừng đăng',
    PublishRecruitment = 'Đăng tuyển',
    UpdateState = 'Cập nhật trạng thái',
    AddNewRank = 'Thêm TTKD',
    UpdateTree = 'Chuyển cây',
    AddNewBranch = 'Thêm nhánh',
    UploadVideo = 'Tải lên Video',
    ChangeLeader = 'Thay đổi đại diện',
    EditTextLink = 'Sửa số text link',
    MAFVerifyInvoice = 'Duyệt hóa đơn',
    MAFVerifyCommission = 'Duyệt hoa hồng',
    MLCompanyAuditingExport = 'Xuất dữ liệu quy hoạch',
    MLCompanyAuditingExportUser = 'Xuất dữ liệu tài khoản',
    M3DLowtour = 'Hạ Tour',
    MRVSeeding = 'Phản hồi nhanh',
    Process = 'Xử lý',
    AddDepartment = 'Thêm mới phòng ban'
}